Actualizacion AgendaDAO-26/12/2020 (casa)
No funciona bien del todo

+++ 100-AgendaDAO/CategoriaFicha.php
@@ -1,10 +1,13 @@
 <?php
-
 require_once "_com/Dao.php";
+require_once "_com/_Varios.php";
+require_once "_com/Utilidades.php";
 
 $conexion = obtenerPdoConexionBD();
 
 $id = (int)$_REQUEST["id"];
+
+$resultado=DAO::categoriaObtenerPorId($id);
 
 $nuevaEntrada = ($id == -1);
 
@@ -14,7 +17,9 @@ if ($nuevaEntrada) {
     $sql = "SELECT nombre FROM categoria WHERE id=?";
 
     $select = $conexion->prepare($sql);
+
     $select->execute([$id]);
+
     $rs = $select->fetchAll();
 
     $categoriaNombre = $rs[0]["nombre"];
@@ -23,14 +28,19 @@ if ($nuevaEntrada) {
 $sql = "SELECT * FROM persona WHERE categoriaId=? ORDER BY nombre";
 
 $select = $conexion->prepare($sql);
-$select->execute([$id]);
+
+$select->execute([$id]); // Array vacío porque la consulta preparada no requiere parámetros.
+
+$rsPersonasDeLaCategoria = $select->fetchAll();
 
 ?>
+
+
 
 <html>
 
 <head>
-    <meta charset="UTF-8">
+    <meta charset='UTF-8'>
 </head>
 
 
@@ -43,35 +53,46 @@ $select->execute([$id]);
     <h1>Ficha de categoría</h1>
 <?php } ?>
 
-<form method="post" action="CategoriaGuardar.php">
+<form method='post' action='CategoriaGuardar.php'>
 
-    <input type="hidden" name="id" value="<?=$id?>" />
+    <input type='hidden' name='id' value='<?=$id?>' />
 
-    <ul>
-        <li>
-            <strong>Nombre: </strong>
-            <input type="text" name="nombre" value="<?=$categoriaNombre?>" />
-        </li>
-    </ul>
+    <label for='nombre'>Nombre</label>
+    <input type='text' name='nombre' value='<?=$categoriaNombre?>' />
+    <br/>
+
+    <br/>
 
     <?php if ($nuevaEntrada) { ?>
-        <input type="submit" name="crear" value="Crear categoría" />
+        <input type='submit' name='crear' value='Crear categoría' />
     <?php } else { ?>
-        <input type="submit" name="guardar" value="Guardar cambios" />
+        <input type='submit' name='guardar' value='Guardar cambios' />
     <?php } ?>
 
 </form>
 
 <br />
 
-<a href="CategoriaEliminar.php?id=<?=$id ?>">Eliminar categoría</a>
+<p>Personas que pertenecen actualmente a la categoría:</p>
+
+<ul>
+    <?php
+    foreach ($rsPersonasDeLaCategoria as $fila) {
+        echo "<li>$fila[nombre] $fila[apellidos]</li>";
+    }
+    ?>
+</ul>
+
+<?php if (!$nuevaEntrada) { ?>
+    <br />
+    <a href='CategoriaEliminar.php?id=<?=$id?>'>Eliminar categoría</a>
+<?php } ?>
 
 <br />
 <br />
 
-<a href="CategoriaListado.php">Volver al listado de categorías.</a>
+<a href='CategoriaListado.php'>Volver al listado de categorías.</a>
 
 </body>
 
 </html>
-
